+++ assets/react/formV2/PrototypeCollectionField.tsx
@@ -1,7 +1,7 @@
 import {Box} from '@mui/material';
 import {FieldArray, FormikValues} from 'formik';
 import React from 'react';
-import {clone} from '../service/Uility';
+import {clone} from '../service/Utility';
 import {StringArrayType} from '../type/Array';
 import {NullableNumberType} from '../type/Scalar';
 import {ElementListType, ElementType, FormBuilder, FormContext, FormFieldRenderPropsType, FormFields, FormFieldValueType, PrototypeCollectionModifiersType, PrototypeCollectionType} from './Form';