+++ assets/react/service/Utility.ts
@@ -1,0 +1,7 @@
+export const clone = <T>(value: T): T => {
+    if ('function' === typeof structuredClone) {
+        return structuredClone(value);
+    }
+
+    return JSON.parse(JSON.stringify(value)) as T;
+};